Use correct namespace for custom phpunit test cases
