+++ UPGRADE-3.x.md
@@ -1,0 +1,6 @@
+UPGRADE 3.x
+===========
+
+## Deprecated AbstractBlockServiceTest class
+
+The `Tests\Block\AbstractBlockServiceTest` class is deprecated. Use `Test\AbstractBlockServiceTestCase` instead.
